Crate forest_bigint

Re-exports

pub use num_integer;

Modules

bigint_ser

biguint_ser

Structs

BigInt
A big signed integer type.

BigUint
A big unsigned integer type.

ParseBigIntError

TryFromBigIntError
The error type returned when a checked conversion regarding big integer fails.

U32Digits
An iterator of u32 digits representation of a BigUint or BigInt, ordered least significant digit first.

U64Digits
An iterator of u64 digits representation of a BigUint or BigInt, ordered least significant digit first.

Enums

Sign
A Sign is a BigInt’s composing element.

Traits

Integer

ToBigInt
A generic trait for converting a value to a BigInt. This may return None when converting from f32 or f64, and will always succeed when converting from any integer or unsigned primitive, or BigUint.

ToBigUint
A generic trait for converting a value to a BigUint.